+++ src/routes/projects/+page.svelte
@@ -18,7 +18,7 @@
     import Feedback from '@components/app/Feedback.svelte';
     import { get } from 'svelte/store';
     import Subheader from '@components/app/Subheader.svelte';
-    import { EDIT_SYMBOL } from '../../parser/Symbols';
+    import { COPY_SYMBOL, EDIT_SYMBOL } from '../../parser/Symbols';
     import AddProject from '@components/app/AddProject.svelte';
 
     const user = getUser();
@@ -71,6 +71,12 @@
             action: (project) => goto(project.getLink(false)),
             label: EDIT_SYMBOL,
         }}
+        copy={{
+            description: $locales.get((l) => l.ui.project.button.duplicate),
+            action: (project) =>
+                goto(Projects.duplicate(project).getLink(false)),
+            label: COPY_SYMBOL,
+        }}
         remove={(project) => {
             return {
                 prompt: $locales.get(
@@ -114,6 +120,7 @@
                     Projects.archiveProject(project.getID(), false),
                 label: '↑🗑️',
             }}
+            copy={false}
             remove={(project) =>
                 $user && project.getOwner() === $user.uid
                     ? {